ddueruem/ddueruem-2021-07/requirements_min.txt
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.3
idna==3.2
python-i18n==0.3.9
PyYAML==5.4.1
requests==2.26.0
termcolor==1.1.0
urllib3==1.26.6
